split builder into sepparated files
